[System]: Use new `WebCompletionSource` instead of `TaskCompletionSource<>`.
This new helper class uses `TaskCompletionSource<>` internally, but instead
of using TrySetException(), it sets a result containing an `ExceptionDispatchInfo`.

The problem with using TrySetException() is that it wraps the exception
object in an `AggregateException`, that would be unobserved if the caller
throws the original exception.

(cherry picked from commit 1e546e5870493d73d4248c3de0f7e6f15ddf6cfc)